requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
astroid==4.0.4
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dill==0.4.1
ecdsa==0.19.1
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
packaging==26.0
passlib==1.7.4
platformdirs==4.5.1
pluggy==1.6.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.4
PyMySQL==1.1.2
pytest==9.0.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
